Update Azure udev rules workaround for WALinuxAgent
For RHEL 8.6, we no longer need the 25rhcos-azure-udev-rules overlay
because the WALinuxAgent-udev package exists. We do still need the glue
code that copies the udev files into the initramfs via a dracut module,
thought. This is not needed for RHEL 9.0.

This migrates the glue code that installs the udev rules into the
initramfs to an overlay module in the RHCOS repo for 8.6. It was no
longer needed in FCOS and removed from the overlay there in [1].

[1] https://github.com/coreos/fedora-coreos-config/pull/1686

From: https://github.com/openshift/os/pull/785